플레이어 테이블 및 플레이어 스텟 컨트롤러 수정
1. 플레이어 테이블 수치 기획서에 맞게 수정
2. 플레이어 스텟 컨트롤러 테이블 값 플레이어에게 적용되도록 수정

+++ EOF/Assets/Scripts/Data/PlayerData.cs
@@ -20,7 +20,7 @@ public class PlayerData
     public float Heal { get; set; }
     public int Action { get; set; }
  
-    public float MaxGauge { get; set; }
+    public int MaxGauge { get; set; }
     public float ComboRate { get; set; }
 
 
@@ -37,7 +37,7 @@ public class PlayerData
         Heal = Convert.ToSingle(row["Heal"]);
 
         Action = Convert.ToInt32(row["Action"]);
-        MaxGauge = Convert.ToSingle(row["MaxGauge"]);
+        MaxGauge = Convert.ToInt32(row["MaxGauge"]);
         ComboRate = Convert.ToSingle(row["ComboRate"]);
